+++ src/sqlfmt/line.py
@@ -87,17 +87,6 @@ class Line:
         prefix = INDENT * (self.depth[0] + self.depth[1])
         return prefix
 
-    @property
-    def has_inline_comment(self) -> bool:
-        """
-        Returns true if the line has an attached comment that was parsed
-        as an inline comment.
-        """
-        if self.comments and self.comments[-1].is_inline:
-            return True
-        else:
-            return False
-
     def render_with_comments(self, max_length: int) -> str:
         """
         Returns a string that represents the properly-formatted Line,
@@ -105,18 +94,17 @@ class Line:
         """
         content = str(self).rstrip()
         rendered_lines: List[str] = []
+        inline_comments: List[str] = []
         for comment in self.comments:
             if comment.is_multiline or comment.is_standalone:
                 rendered_lines.append(
                     comment.render_standalone(max_length=max_length, prefix=self.prefix)
                 )
             else:
-                # comment is potentially an inline comment, and we'll handle
-                # that below. Inline comments must be the last comment
-                pass
+                inline_comments.append(comment.render_inline())
 
-        if self.has_inline_comment:
-            rendered_lines.append(f"{content}{self.comments[-1].render_inline()}")
+        if inline_comments:
+            rendered_lines.append(f"{content}  {' '.join(inline_comments)}")
         else:
             rendered_lines.append(f"{self}")